Move WireMock to UITestsFoundation
