+++ WordPress/UITestsFoundation/BaseScreen.swift
@@ -45,3 +45,24 @@ open class BaseScreen {
         }
     }
 }
+
+// MARK: - Dump of files from the other targets
+// All in this one to avoid messing up with the project file during the transition...
+
+import Foundation
+
+class WireMock {
+    private static let hostInfoPlistKey = "WIREMOCK_HOST"
+    private static let portInfoPlistKey = "WIREMOCK_PORT"
+
+    static func URL() -> Foundation.URL {
+        let host = infoPlistEntry(key: hostInfoPlistKey)
+        let port = infoPlistEntry(key: portInfoPlistKey)
+        return Foundation.URL(string: "http://\(host):\(port)/")!
+    }
+
+    private static func infoPlistEntry(key: String) -> String {
+        let plistUrl = Bundle(for: WireMock.self).url(forResource: "Info", withExtension: "plist")!
+        return NSDictionary(contentsOf: plistUrl)![key] as! String
+    }
+}
